Нововведения и рефакторинг - обновлено ядро WinForms - практически закончены работы по сборке проектов - реализован почти полный функционал компиляции проектов - модифицированы и улучшены алгоритмы получения свойств объектов редакторов - добавлена архитектура ресурсов VRSF - Void ReSource Files - реализован тестовый алгоритм сборки кода в PHP варианте (не VLF) - обновлены константы (STUDIO_DIR -> APP_DIR) - добавлено окно сборки (компиляции) проектов - добавлен список существующих на форме объектов - проведены небольшие оптимизационные работы над "main.vlf" - исправлен счётчик времени запуска среды - добавлено сохранение путей сборки / компиляции проектов - немного обновлён алгоритм "VoidEngine\VoidEngine::export(import)Object" - предупреждены ошибки в работе "VoidEngine\EngineAdditions::getProperty" - добавлено закрытие программы в метод "$APPLICATION->restart ()" (для параноиков~) - добавлена самая полезная функция движка: "VoidEngine\nothing". Не делает ничего. Вообще. - всем спискам добавлено дополнительное свойство "names", которое возвращает массив названий элементов списка - оптимизирован код "VoidEngine\VoidDesigner" - в "VoidEngine\VoidDesigner" портирована функция ядра "AddComponent" - исправлена глупая ошибка метода "getComponentClass" компонента выше - "VoidEngine\VLFInterpreter" добавлено свойство "throw_errors", отвечающее за вывод ошибок интерпретации VLF разметки

+++ system/core/VoidEngine.php
@@ -13,7 +13,7 @@
  * @see         license.txt for details
  * @author      Podvirnyy Nikita (KRypt0n_) & Andrey Kusov
  * 
- * @version     build-2019/02/02
+ * @version     build-2019/02/03 (state-y/m/d)
  * 
  * Contacts:
  *
@@ -42,7 +42,7 @@ $GLOBALS['__debug'] = [
     'error_count' => 0
 ];
 
-if (is_array ($files = scandir ('debug')))
+if (is_dir ('debug') && is_array ($files = scandir ('debug')))
     foreach ($files as $id => $file)
         if (is_file ('debug/'. $file))
             unlink ('debug/'. $file);
@@ -54,7 +54,7 @@ require 'common/Globals.php';
 require 'common/Constants.php';
 // require 'common/LanguageFiles.php'; // Признан устаревшим
 
-if (is_array ($exts = scandir ('extensions')))
+if (is_dir ('extensions') && is_array ($exts = scandir ('extensions')))
     foreach ($exts as $id => $ext)
         if (is_dir ('extensions/'. $ext) && file_exists ($ext = 'extensions/'. $ext .'/main.php'))
             require $ext;
